// apps/api/generated/server/model/createAccountRequest.ts
/**
 * QZD Platform API
 * The QZD Platform API exposes authentication, account, ledger, remittance, and validator administration capabilities for wallet operators and platform integrations. The contract is defined up-front to enable client/server scaffolding via OpenAPI Generator. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateAccountRequest { 
    ownerId: string;
    /**
     * ISO 4217 currency code for the account.
     */
    currency?: string;
    tags?: Array<string>;
    metadata?: { [key: string]: string; };
}