// apps/api/generated/server/model/ledgerEntry.ts
/**
 * QZD Platform API
 * The QZD Platform API exposes authentication, account, ledger, remittance, and validator administration capabilities for wallet operators and platform integrations. The contract is defined up-front to enable client/server scaffolding via OpenAPI Generator. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LedgerEntry { 
    id: string;
    type: LedgerEntry.TypeEnum;
    amount: string;
    currency: string;
    direction: LedgerEntry.DirectionEnum;
    counterpartyAccountId?: string;
    memo?: string;
    reference?: string;
    createdAt: string;
}
export namespace LedgerEntry {
    export type TypeEnum = 'transfer' | 'issuance' | 'redemption' | 'remittance';
    export const TypeEnum = {
        Transfer: 'transfer' as TypeEnum,
        Issuance: 'issuance' as TypeEnum,
        Redemption: 'redemption' as TypeEnum,
        Remittance: 'remittance' as TypeEnum
    };
    export type DirectionEnum = 'credit' | 'debit';
    export const DirectionEnum = {
        Credit: 'credit' as DirectionEnum,
        Debit: 'debit' as DirectionEnum
    };
}